change `tsfeatures` to work around a Numba bug on Python 3.7
Summary:
GitHub tests were failing on Python 3.7 but not 3.8 or 3.9. Much debugging showed that `get_statistics` was not returning all the features passed in through `dict_features` but only the first one. Removing the use of `partial` and other changes did not fix this. Removing `jit` decorator did fix this.

Also
* change `self.get_X` to `TsFeatures.get_X` for all static methods.
* use `functools.partial` only where needed

Reviewed By: jeffhandl

Differential Revision: D29726351

fbshipit-source-id: 5534714c288884feb81836c120f0e69ea6c3a6a4