Adding a time delay metric to the TuringEvaluator
Summary:
Add a delay metric in the measure function.
Here the delay is computed similarly as the Recall, by first computing the delay for each annotator, and then averaging all delays.

Another possible way, is to group the change points from all annotators, and to take the best (earliest) as the "ground truth" before computing the delay of our algorithm.

Also formatted the test_detectors.py

Reviewed By: souravc83

Differential Revision: D29890251

fbshipit-source-id: 44e11400897cfe2eff8c8a4d7e9d9be3e9bece57