adding data loading
Summary:
This is 1/n diffs to add data loading functionalities and standardize the way to load data across Kats, especially for our unit tests.

We convert the `kats/data/` directory to a sub-module and implement our data loaders in `utils.py`, so for a specific data we can load it through `load_data_name()` method.

This diff is an attempt for `air_passengers` data, we will address other data sets in future diffs.

Reviewed By: michaelbrundage

Differential Revision: D30523794

fbshipit-source-id: 00675cf6c5b6815f23cc8d6431ec66bdc923faa1