Add .to_numpy() to Detector plot methods
Summary:
CUSUMDetector and RobustStatDetector are currently trying to plot pandas columns directly which is leading to the below error.

Here we add .to_numpy() suffix to the columns to clear the error.
This now runs correctly as below.

The .values suffix has already been implemented in some other classes (e.g. BOCPDetector) , which we update to .to_numpy() for consistency.

Differential Revision: D30765628

fbshipit-source-id: 0f37180be71c271887080f43e62c03ac73bafa35