Add .to_numpy() to Detector plot methods (#123)
Summary:
Pull Request resolved: https://github.com/facebookresearch/Kats/pull/123

CUSUMDetector and RobustStatDetector are currently trying to plot pandas columns directly which is leading to the below error.

Here we add .to_numpy() suffix to the columns to clear the error.
This now runs correctly as below.

The .values suffix has already been implemented in some other classes (e.g. BOCPDetector) , which we update to .to_numpy() for consistency.

Differential Revision: D30765628

fbshipit-source-id: 3dcdc021e97c794e2f46e84edd378fa7ee9aa5f1