Force Bayesian Var Model to only accept multivariate time series
Summary:
The issue here (https://github.com/facebookresearch/Kats/issues/127) was caused by the bayesian var model being initialized with a univariate time series. After consulting with rakshita95, we determined that it made sense to limit the scope of the model to only multivariate time series. This diff makes that change and raises a `ValueError` if a univariate time series is passed to the model. It also includes an updated test for this case.

Also noticed while working on this diff that the Bayesian Var model is not compatible with `TimeSeriesData` objects that have a time column name other than 'time'. Fixed this bug and added a relevant test.

Reviewed By: jeffhandl

Differential Revision: D31194715

fbshipit-source-id: e92bbcd3462231f41a68df01f70cea28eec9432a