Fix Pyre errors in models/lstm.py
Summary:
Fix all but one Pyre error in `models/lstm.py`. The remaining one is because `predict()` signature doesn't match the superclass, which is a larger issue to address.

This appears to have revealed an existing latent bug. `predict` was setting `hidden` instead of `hidden_cell` on the model.

Also improve documentation formatting while I'm there.

Reviewed By: rohanfb

Differential Revision: D31621538

fbshipit-source-id: 7be53a46bc744ec8da4d5cfaf3de6d090eecaf12