Improved Efficiency of test_ensemble.py [1/N]
Summary:
This diff is focused on the test cases for `BaseEnsemble`.  Subsequent diffs will cover the rest of `test_ensemble.py`.

I wrote the function `get_fake_preds` to generate a fake forecast output using `np.random`.  For the purposes of these test cases, the actual values of the forecast predictions don't seem to matter, so we can use this function to mock the outputs of each of the individual models.

Reviewed By: rohanfb

Differential Revision: D31651984

fbshipit-source-id: 285641463aaa345a5928ec850a701d1cdb0af07c