Fix pyre errors and increase code coverage for `detectors/stat_sig_detector.py`
Summary:
Fix most Pyre type errors and increase code coverage to 100%.

Remove `to_datetime`, which wasn't used and couldn't work anyway (it attempts to divide a datetime by a float).

`_set_unit_time` had a comment that it first checks historical data, but that's the opposite of what the code does, and in fact the code was throwing away the result of `infer_freq` on the historical data. So, remove the comment and fall back on historical data for the time unit if the data doesn't already have one.

`_handle_not_enough_history` had a check on `time_unit` being None that couldn't be reached, because `_should_update` is called before `_handle_not_enough_history` and asserts if `time_unit` is None. So, change to an assert and remove the duplicative call to `_should_update`.

Reviewed By: rohanfb

Differential Revision: D31763564

fbshipit-source-id: fdf12305e286b2f4fc312cec7af209283dfa73f1