Fix NaN bug in Outlier Detector
Summary:
Currently, the output of Outlier detector has a bug, that the first 3 and the last 3 values are NaN.
You can see this in N1331329.

The reason is, when STL decomposition is run, the trend is not calculated for the first 3 and last 3 points by default.
By using the extrapolate argument, as described in the doc, we fix this.
https://www.statsmodels.org/dev/generated/statsmodels.tsa.seasonal.seasonal_decompose.html

Reviewed By: yangbk560

Differential Revision: D32297817

fbshipit-source-id: 94d0e7d01c4e36aafd7c935209b51281c0f4c5f5