Add AutomaticForecastingGP model to KATS
Summary:
This adds the AutomaticForecastingGP model to KATS by creating a wrapper around it using the `kats.Model` API. The model uses a lot of infrastructure from the `beanmachine.facebook.timeseries` library which is not open-source currently. As such, I have added this to `kats.internal` and we can plan to make it publicly accessible when the original code is open sourced.

For the time being, this will let us expose this model for internal use via the `kats` bento kernel and for forecasting collaborations, e.g. ALM that we have on our roadmap. In the future, we may want to deploy this model to forecaster as well as expose other models in this module that we are currently working on.

Questions for reviewers:
 - I have added a test to `gp_model/test/test_autoforecast_daily.py` since there wasn't an internal directory in `tests`. This seems to run fine but let me know if this should be moved elsewhere.
 - The AutomaticForecastingGP model can take in covariates which can be included in the `TimeSeriesData`, i.e. the response may depend on factors other than time. This isn't something I saw in the other models, but I have tested this and it works fine.
 - Please surface any questions/concerns you may have.

Reviewed By: yangbk560

Differential Revision: D32571170

fbshipit-source-id: 90c537f9b6e65d6887cad5c9d68d56ba2845c72f