Improve implementation of extra regressors
Summary:
Current implementation of Kats prophet for multi-regressors is not ideal. For example, one can specify regressors without corresponding timestamps, which is prone to errors. And provided that `TimeSeriesData` handles multivariate time series data, it's more nature to include the regressor information in the `TimeSeriesData` objects. In addition, the `.predict()` method is also modified accordingly.

Main changes:
1) Moving regressors info (i.e., regressor values) to `data` and allowing `ProphetModel` to accept multi-variate time series.
2) For `predict` function, `future` is required for model with regressors. In addition, when `future` is not None, it's length for future dates should be no less than `steps`.

Next steps:
1) implementation of `cap` and `floor` is limited. i.e., fbprophet accepts non-constant `cap` and `floor`. And when `growth=linear`, `cap` will have impact on the PI, which is not correctly implemented in Kats.

Reviewed By: rohanfb

Differential Revision: D35268781

fbshipit-source-id: 1f404706182f4359db268ae60ca9e3ca415f5834