Bugfix to TimeSeriesData constructor for multivariate time series
Summary:
When constructing a TimeSeriesData, Kats checks the dtypes of the Series or DataFrame to ensure they are numeric. However, for multivariate data with a nonnumeric dtype (sometimes encountered when passed an empty DataFrame), Kats fails with the following exception:
```
Traceback (most recent call last):
...
  File "...#link-tree/kats/consts.py", line 330, in __init__
    logging.warning(self.value.dtype)
  File "...#link-tree/pandas/core/generic.py", line 5274, in __getattr__
    return object.__getattribute__(self, name)
AttributeError: 'DataFrame' object has no attribute 'dtype'
```
This is because DataFrames have no dtype (need to use dtypes).

In fixing this, uncovered that the unit test for initialization errors was only checking the first error case. Upon fixing that unit test, uncovered that checking for mismatched time and value lengths was broken. That bug will be fixed in a subsequent diff.

Reviewed By: jeffhandl

Differential Revision: D36749007

fbshipit-source-id: a33a2f9ecc476aadd69f26da6e07d1ecb2c0cf7a