Bugfix to TimeSeriesData constructor validation of time and value lengths
Summary:
According to the unit tests, if `time` and `value` are explicitly passed to the TimeSeriesData constructor, a ValueError should be raised if the time and value lengths mismatch. However, when the lengths are mismatched, `TimeSeriesData.to_dataframe` ends up filling missing data with NaNs, masking this issue.

To correct the issue, this diff moves the length validation up before this method is called for the one case where there could be mismatched lengths. It also adds back the unit test that checks for this.

Note that as a result of this bug, the implicit API for Kats has been that mismatched lengths will be filled with NaNs. This diff could break existing code that relies on this behavior.

Reviewed By: jeffhandl

Differential Revision: D36749021

fbshipit-source-id: 69056bc885f78a55f89f486363e0baf6678f6935