Added Base Embedding Model
Summary: This is the base class that will be used for our embedding models going forward.  After this diff is shipped, we will update `T2VNN`, `VAE`, and `AE` to inherit `EmbeddingModel` and their parameters to inherit `EmbeddingParams`.  We will also ensure that all subsequent embedding models use this API.

Reviewed By: peiyizhang

Differential Revision: D37184823

fbshipit-source-id: a03364f91a566c6bcf3b2b6abc104b75b5809e25