Added Minimum Percentage Change
Summary:
Added a minimum percentage change parameter to the stat sig algorithm. This is to reduce false positives for the ads feature serving team (but will potentially help others as well).
Even though a change. is statistically significant, the scores will be zero, if the change does not cross a % change threshold.
So, a change in mean from 0.997 to 0.995 is meaningless, even it is statsig, because the time series has very small variance.

Reviewed By: peiyizhang

Differential Revision: D38958157

fbshipit-source-id: 34a013074cbe4bed19afc2531d37f3244935ad1b