Report column name and its type when failing type check
Summary:
When type check failed for DF, it is helpful to report the column and its type instead of just types. Before this change, you will receive error message like `Time series data is type [dtype('float64'), StringDtype] but must be numeric` which is hard to figure out which exact column is of the wrong type. With this change, the error message becomes:
```
ValueError: Time series data is type {'value': dtype('float64'), 'ds': dtype('O')} but must be numeric
```

Reviewed By: yangbk560

Differential Revision: D39120381

fbshipit-source-id: 5ebcfe38b088b59ff5a794a2dcf34dea0833d1ee