Unbreak `test_metalearner` unit tests
Summary:
In D39736772, I misinterpreted whether my diff broke the Kats tests –– I thought the breakage was unrelated, but it wasn't.

This diff should fix the tests, but one question is whether we actually want to change the tests instead? I'm not sure if you were expecting to see the generators suggest duplicate points here; if not, we might want to change the tests rather than apply this change. Let me know what you think!

In any case, feel free to land this and we can figure out the deduplications later –– in the meantime, this aligns the current behavior with behavior before D39736772 for the purpose of unbreaking the tests.

Reviewed By: souravc83

Differential Revision: D40083294

fbshipit-source-id: a4edb500bf31d274bb233c04223c055f5723cbfe