Allowing explicit vectorized flag in CUSUM to override existing value
Summary: CUSUMDetectorModel currently ignores the `vectorized` flag (defaulting to False) if a serialized model is provided. This diff updates CUSUMDetectorModel to always use the provided value for vectorized over anything in the serialized model while still defaulting to False. This covers cases where old serialized models are provided but vectorized is set to True.

Reviewed By: peiyizhang

Differential Revision: D42334868

fbshipit-source-id: 5c3cf32b71bfa62fc66cd84255be69adb3342149