Abstract Null Distribution with Dependence
Summary:
**diff adds functionality, but defaults all choices to current behavior (with no dependency), no actual changes will occur in prod detectors**

Add dependency to the null distribution. Here is some motivation of why this is important:

Assume the test statistic was:
```
x_t = N(\mu_t, 1) for odd t
x_t = x_{t-1} for even t
```
In this case, when we test `H_t` and `H_{t+1}` when t is even, the conclusion is the same for each test. In this case, the chances of seeing a run of `d` rejections is much higher than when we assume independence in the test statistic. We can account for the increased chances of seeing a run of rejections by adding dependency (correlations) to the test statistic.

Changes also include:
- Breaking apart `_vec_solve` into modular pieces
- Creating autocorrelation matrices from `ARMA(p, q)` processes
- Creating different implementations of `_w`.

Differential Revision: D42080487

fbshipit-source-id: c6daa251cb18cfefa04fb028d0961d6160f1b351