Conditional seasonality was added to the prophet detector to kats.
Summary:
Conditional seasonality was added to the prophet detector to kats. In addition, it's now possible to provide an argument to the prophet throw ProphetDetectorModel (via prophet_arg dict). And it's also now possible to provide seasonalities as a list of dict via seasonalities arguments.
In addition, it's now allowed to use a multicolumn dataframe in ProphetDetectorModel. One of the columns should be named y (prophet value column name) or value (Kats value columns name).
All these changes needed to make conditional seasonality available in ProphetDetecotModel

Reviewed By: uthakore

Differential Revision: D43126105

fbshipit-source-id: e0367f295d3a5a982da196df3c5bc12436d0c855