Rename density distance model --> distribution distance model
Summary:
DistributionDistanceModel is a more apt and immediately understandable name for the algorithm since it takes a CDF rather than a PDF as input and is primarily intended to be used for distribution distance-based detection.

Also fixing some minor bugs, removing superfluous params, renaming other params for clarity, and improving error specificity.

Reviewed By: peiyizhang

Differential Revision: D46012453

fbshipit-source-id: cb9e684c3d601011ff7f92693278986c9268d3c8