Make expanding window behavior explicit for RollingStats
Summary:
Currently, when insufficient historical data is provided (historical data length < rolling_window in points), the algorithm would use as much data as is available (expanding window) to compute stats. This is more robust to corner cases, but can be misleading if the user only wants precise stats. This diff makes that behavior explicit and raises a DataError if there isn't enough data.

This is an issue for 1D when the rolling_window is set very large, but enough data is not present in source and/or not fetched at priming time.

Note that this only applies for point-based rolling_windows because missing data is indistinguishable from irregular granularity data, so with time-based windows, there can legitimately be different numbers of points per analysis window. In the time-based rolling_window case, we rely on users to pay attention to backtesting results and not set nonsensical rolling_windows.

Reviewed By: peiyizhang

Differential Revision: D45836407

fbshipit-source-id: b4c0e4478f7fb568106d6d5ae3adcecb67d0590b