Fixing OutlierDetectorModel
Summary:
OutlierDetectorModel does not obey the contract that all other DetectorModels do (all series in the AnomalyResponse returned by `fit_predict()` and `predict()` have the same time values; those time values are the same as in the `data` that was passed in). This is caused by:
1) a bug in how historical data is handled in `fit_predict()`
2) another bug in `predict()` where the `predicted_ts` can have more datapoints than the input `data` (caused by interpolation).

This diff fixes the bugs and adds some unit test cases around this.

Reviewed By: peiyizhang

Differential Revision: D46014354

fbshipit-source-id: 2e3698da8c85082989771d10a2fb53ac025c1684