replace np.float with float [3]
Summary:
`numpy.float` has been deprecated and removed in numpy-1.20.0. It was an alias to the builtin `float` and should be safe to replace with `float`.

The whole change was mechanical with the following one-liner:
```
fbgr -sl '\bnp\.float\b' | xargs perl -pi -e 's,\bnp\.float\b,float,g'
```

Differential Revision: D46609269

fbshipit-source-id: 7ece32e764de0da51636f772349f1ede587fb400