Support for mutable object for timeseries tuning.
Summary:
During optimization we are using multithreading, to that we need to copy evaluating function across the threads. Because otherwise it may cause problem. Therefore, evaluation function needs to be serializable. If evaluating function is a class function, the whole class will be serialized (because there are some object properties used in function), and if class has a socket as a property, it may cause an issue during serialization.
But some objects can be mutable instead of being seriliazeble, that meen that we can copy that across threads without deepcopy.
arc lint also added minor changes in Target file

**Internal**
 This commit should adress for these type of objects.
according to this user problem
(link in the comments as it description going to github)

Differential Revision: D48608190

fbshipit-source-id: 190a525e01569b0c8a6962684854ff8985ee7be5