Providing support for fixed parameter with nevergrad, defined in ax-like
Summary:
as occasuianately convert "choice" parameter to "fixed" if there one value in the list.
Nevergrad didn't support that type, so I've provided a conversion.
We use partial function to set it up instead of list of one element in order to reduce dimensions count.
this is important to do, becaouse nevergrad doesn't works good with huge dimensions.

Differential Revision: D49957021

fbshipit-source-id: 1bb8451cd9fbfcada9b7a05b82bb9ed75295bee7