Adding changein_direction parameter setup in differen ways
Summary:
We are adding abilities to setup both direction in cusum not only as None, but also as "both" or empty string or list. It is the same by logic and will be used when passing this value throw thrift as string. I've also added ability to setup by list or string. This maybe helpful.
No actual changes in algorithm or somewere.
I've also modified test to use both old and new type of parameter.

Differential Revision: D48919825

fbshipit-source-id: 4b730b815848ab22e995d407058c78e504a45065