support of seasonalities as list of string
Summary:
Currently, prophet defines a custom enum for seasonalities. There is no proper way to expose this enum to configerator to validate seasonality param at authoring time.

We can add a new thrift structure, but it will add maintenance overhead as we need to keep mapping between thrift structure and the enum.

As any python enum class expose the enum as a key, we can use this to have an implicit mapping between gp param and kats enum.

This diff enable user to use the seasonality name and the code will convert it to enum during initialization

Differential Revision: D51238962

fbshipit-source-id: 5f35f8d8ffcfe3db6b498d1bcdb460004e76c579