adding helper function to aggregate df against intervals
Summary:
with this funciton we apply interval list to some dataframe to apply aggregation on thesee interval.
for example we want to get sum of values of some timeseries on specific intervals.
This helper function will be used for metric customization in autotuning. As we got request from Ads to create a target function which measure sum of revenue drops.
But in this diff it's the only function without any usage and unit test to be sure, that it works correctly.

Differential Revision: D51549538

fbshipit-source-id: 2eaf43da91784d65749a5d80119d9706d1cf6a4b