Support fixed parameters for autotuning without additional settings
Summary:
This diff address parameters dimensional problem.
For optimizers like nevergrad parameters space dimension size is the important argument.
But in fact some parameters has only one possible value and can't be optimized. We usually reducing dimensions by removing these parameters and setting is as default.
But because of nevergrad interface we couldn't pass fixed paramters to optimizer. So we used to have a flag fixed_params_in_space to pass them to optimizer.
Unfortunately if this flag setting to false, we fault with creating detectors, which expected to recieve these fixed parameters. For example history window for the cusum detector.
For these detectors users have to setup this flag explicit, which to True, which is not user friendly.

With fxing this issue, fixed parameters will be propagated for nevergrad optimizer as function parameters by partial. And user will be able to use fixed parameters in optimization

Differential Revision: D53524719

fbshipit-source-id: 49d42421a0c1d8f09e3369d3d79a933549f4a5f4