Moving detectors.utils unit test internally
Summary: The detectors.utils module is an internal file not synced to GitHub. It is tested by tests/detectors/test_utils.py, which is synced to GitHub. Becuase the detectors/utils.py file is kept internally the test fails to import successfully, causing the entire CI Actions workflow to fail. This diff moves the unit test to tests/internal (which isn't synced to GH) and updates the TARGETS

Differential Revision: D53367618

fbshipit-source-id: a0f5ca1fae9c85d4e70ca1d0654b30533f44a800