Limiting country list by one country similar to prophet
Summary:
In the Prophet implementation, only one country is supported for holidays. Although you can call add_country_holiday multiple times, only the last country specified will be used. This means that if a list of countries is provided, only the last one in the list will be considered.
Due to this limitation, it is necessary to restrict the number of countries to one, as implemented in Prophet. With this fix, only one country can be provided at a time.
Here's the link to the GitHub repository where the logic is implemented: https://github.com/facebook/prophet/blob/c00f6a2d72229faa6acee8292bc01e14f16f599c/python/prophet/forecaster.py#L744

Differential Revision: D55373227

fbshipit-source-id: 16f876155950ad38f1f4d04bf045595153b0300c