fix slow tests for metalearner
Summary:
Certain kats tests are disabled, since they are timing out and failing. This diff improves the performance of the tests by removing some of the slow features being computed.

More precisely, this diff removes `"firstmin_ac"` feature. It also propagates the params to use only selected features, such that this functionality may be reused when creating models.

Analysis of slow features was performed in N5487518.

Differential Revision: D58581652

fbshipit-source-id: ad846aad9cdd5c8b0994339ac2a3308d111bcc50