UBN tests
Summary:
A bunch of tests are flaky or failing. This seems to be holding up an FBPkg release (which fired a UBN), and then is also just creating a lot of tasks for flaky or failing tests.

This diff was originally used to push to py sci stack v2, and update all the broken/tests and code, however I've since split it into that v2 upgrade (next diff in stack) and this smaller diff which just resolves already-broken tests.

Failing tests have two kinds of causes: test code is broken, or test is super heavy weight and timing out always. Flaky tests seem to just be heavyweight, and their success seems to depend on whether they get enough resources to finish in 10 min or not.

First category I just fix the test or correct the code (usually some API changes in py 3.8 ).

Second category is a mixed bag: where possible I rewrote the test to mock out the expensive calls (I'm looking at you TsFeatures, which in some tests was called nearly 10 times). If mocking out the expensive part seemed infeasible, or just sufficiently daunting, I labelled the test as `heavyweight` and `long_running`, which uses more resources but hopefully stops timing out. This only ended up happening for the flaky tests, which were on the line of not timing out, so hopefully this fix should do the trick.

#buildall

Reviewed By: irumata, igorsugak

Differential Revision: D58582196

fbshipit-source-id: 04d7fec197747237a24f953c274333972a02e820