Fix bug in cusum prediction
Summary:
We encountered an issue with the CUSUM prediction when we generated only one chunk of prediction because we were attempting to extract frequency information. The CUSUM method reorganizes data into chunks and then reconstructs it based on frequency, but if we only have one chunk, there is no frequency information to extract.
In reality, we can obtain frequency information from the data, but we lose it after reconstruction. Reconstruction is necessary to split the data into chunks for prediction purposes.
To address this issue, we provided a fix that generates a separate data point using the same logic.
The problem has a blast radius of 637 requests per day (as seen in DoD and detector gadget).
Here's an example request and response: https://fburl.com/thrift_fiddle/8tnn9q1a
The issue we're facing is that we're not receiving the proper message here.

The issue only occurs for the last chunk when backtesting or creating a model, so it doesn't completely block users. However, it does provide noise for us and slightly affects our metrics.
To address this issue, I obtained the frequency from historical data for this specific case, and it is now working correctly.
I kept the old logic for all other cases to reduce the blast radius of the diff (to only failed cases) in case something goes wrong.

this example request and responce https://fburl.com/thrift_fiddle/8tnn9q1a

the problem we don't getting proper message here

IMPORTANT we still will have problem if predicted data provided empty. Need to invest more to identify the correct behaivour there.

Reviewed By: islijepcevic

Differential Revision: D63729249

fbshipit-source-id: d7bb4300342919b201639363fcdd31c7bf9b1cc3