temporarily disable test_tsfeatures
Summary:
This test is failing in LLVM17 for firstmin_ac.

After further inspection and running that tsfeatur step by step, it looks like LLVM17 returns the correct value, while the current code and test is incorrect.

After D61452958 lands, we should be safe to update the expected value of firstmin_ac to 8 and enable the test, which this diff already does.

Reviewed By: meyering

Differential Revision: D64551795

fbshipit-source-id: 56955458747423b41e28a41f3d6f20aecc1861f6