remove outliers based on margin
Summary:
Prior to D67392638, Prophet Detector was removing all data as outliers when the incoming data came at irregular intervals, and thus causing errors during training.

D67392638 fixed the issue by reverting the outlier removal if it removed all data.

Here we change our approach how we fix the issue. Instead of reverting the outlier removal, we expand the confidence band of the outlier removal procedure with a small margin. This way, we still are able to remove eggregious outlires, but we don't run into the issue of removing all data.

Reviewed By: yazovskymeta

Differential Revision: D71908123

fbshipit-source-id: 4f03aeeb9f663c5911752f654dbcad9ee629abac