Remove unused numpy import in kats/models/var.py
Summary:
## Instructions about RACER Diffs:
This diff is removing the unused imports from Code Quality Score from [Monetization codehub](https://fburl.com/quality/58ozwacb):
 - Accept and ship the diff (racer do not do auto land)
 - [**Recommended**]Commandeer and land this diff.

If you are happy with the changes, please accept and ship, or commandeer it if minor edits are needed.
If you are not happy with the changes, please comment on the diff with clear actions and send it back to the author. Racer will pick it up and re-generage.
If you really feel the Racer is not helping with this change (alas, some complex changes are hard for AI) feel free to abandon this diff.

This diff was pre-created by Racer AI agent for your convenience on top of T226646276. More details about the program see this [post](https://fb.workplace.com/groups/1177219810659335/permalink/1198503531864296/). For questions or suggestions please post in [RACER Feedback and Q&A](https://fb.workplace.com/groups/2477474979269093) group.
## Summary:
Removed unused 'numpy as np' import from fbcode/kats/kats/models/var.py as identified by Code Quality Score.
 ---
> Generated by [RACER](https://www.internalfb.com/wiki/RACER_(Risk-Aware_Code_Editing_and_Refactoring)/), powered by [Confucius](https://www.internalfb.com/wiki/Confucius/Analect/Shared_Analects/Confucius_Code_Assist_(CCA)/)
[Session](https://www.internalfb.com/confucius?session_id=f5ff6030-429d-11f0-8cbe-eac49f32f9a9&tab=Chat), [Trace](https://www.internalfb.com/confucius?session_id=f5ff6030-429d-11f0-8cbe-eac49f32f9a9&tab=Trace)

Reviewed By: proof-by-accident

Differential Revision: D76109462

fbshipit-source-id: 31f32c832c4844117373eaed231578027a1a58ab