Move the df renaming logic to standalone util
Summary:
This is a generally useful utility for multivariate algos where we'd want to map column names in the input dataframe to more canonical names based on the underlying algo.

For example, a df from rapido -> 1d -> ML stack might have a df of the following format where only the entity prefix upto `@#$` may be known ahead of time (the suffix might something dynamically computed)

| time   | foo@#$123  | bar@#$123   |
| -- |
| T  | 1   | 2  |
|  T | 2   | 3  |

We would want to transform this into something semantically meaningful

| time   | control | test  |
| -- |
| T  | 1   | 2  |
|  T | 2   | 3  |

In order to do this, the caller has to at least provide the prefix of the input cols so we know how to map them to a standard format.

Moving this out into a standalone util

Differential Revision: D84464586

fbshipit-source-id: d011444475d6d1f4db490885239b9c2628ffc1e3