New impl for rename_columns_by_prefix
Summary:
Currently the `rename_columns_by_prefix` function is a bit too strict: it required every column (except "time") to be matched by a prefix in the mapping, and would throw an error if any column wasn’t covered. That meant we couldn’t have any extra columns in the input df that didn’t fit the prefix rules.

Now, we’ve relaxed it: the function just renames the columns that match a prefix, and leaves everything else alone. The only strictness that remains is that every prefix in the  mapping has to match exactly one column, and no column can be matched by more than one prefix (to avoid ambiguity).

Differential Revision: D84502415

fbshipit-source-id: 39055bc6983d916ec7f512a133ffde65b689c274