PSS3 fixes using ruff command
Summary:
Python foundation team has provided a command to help migrate known issues between PSS2 and PSS3.

All changes in this diff were generated by running this command:
```
cd ~/fbsource
./fbcode/tp_foundation/codemods/pss/ruff/ruff --config fbcode/tp_foundation/codemods/pss/ruff.toml check --fix fbcode/kats
```

I have verified that all of these changes are safe and do not change logic of the code. They are mainly renamed API calls in the new versions of numpy.

More details in the [post](https://fburl.com/workplace/5x000var)

Note that as of this diff, kats is still in PSS2.

Differential Revision: D84500868

fbshipit-source-id: 5f67c3840c205a8048f67ad99388528f5833e6be