add tests for _mvnun
Summary:
[mvnun](https://fburl.com/wkosthho) is a subroutine implemented in Fortran for computing multivariate normal distribution function values. Our wrapper `_mvn_mvnun` implements integrals on the  multivariate normal function, which is then used within `IntervalDetector`.

`scipy.stats.mvn.mvnun` has long been deprecated, and it will not be supported in PSS3. According to [github issue](https://fburl.com/ed3bgar3), this funtionality was never supposed to be public in the first place.

Here we add tests for our usage of `mvnun`, which can help us migrate in the follow-up diffs.

Note that as of this diff, kats is still PSS2

Reviewed By: seunghaklee

Differential Revision: D84516264

fbshipit-source-id: ea425d6fceb17342bf7a196ab2666f846a7f1e60