Return error results per timepoint in case of error
Summary:
Inputs with invalid values (like the `comparison_treatment_denominator`) were resulting in the deltoid algo throwing an exception and returning an error for the whole evalaution.

That is not correct, since other timepoints can have the correct data. Instead, this change drops the points for which there is an exception.

This also changes how the debug informations are propagated, so that we now have the debug informations for every point that is provided in the input.

There are a few more changes to utilities that weren't properly considering that there is no guarantee on the name of the time column.

Differential Revision: D86196957

fbshipit-source-id: e128a69fa940154173732f1326ccb5c62b2ee343