fix or ignore type annotations that fail in PSS3
Summary:
Adding type annotations where possible, or ignore comments in other places.

Certain methods, such as `pd.to_datetime`, can return different types, depending on the input. Given that this is poorly annotated in the core libraries (such as pandas), pyre has trouble following the types that are actually used. For example, pyre would think we are using `pd.Timestamp`, while we are actually operating with `pd.DateTimeIndex` and our code would work.

This diff opted to add `pyre-fixme` annotations in the code paths that are not critical, while enforcing the stricter type annotations in code paths that are more critical.

Note that as of this diff, kats is still in PSS2.

Differential Revision: D84603370

fbshipit-source-id: 0ce17ea160aa1eb419150487f54a8bb13bfcfd79